+++ web/src/pages/CreateOrphanage.tsx
@@ -1,19 +1,11 @@
 import React from "react";
 import { Map, Marker, TileLayer } from 'react-leaflet';
-import L from 'leaflet';
 import { FiPlus } from "react-icons/fi";
 
 import '../styles/pages/create-orphanage.scss';
-import mapMarkerImg from '../images/map-marker.svg';
 import Sidebar from "../components/Sidebar";
+import mapIcon from "../utils/mapIcon";
 
-const happyMapIcon = L.icon({
-  iconUrl: mapMarkerImg,
-
-  iconSize: [58, 68],
-  iconAnchor: [29, 68],
-  popupAnchor: [0, -60]
-})
 
 export default function CreateOrphanage() {
 
@@ -36,7 +28,7 @@ export default function CreateOrphanage() {
                 url={`https://api.mapbox.com/styles/v1/mapbox/light-v10/tiles/256/{z}/{x}/{y}@2x?access_token=${process.env.REACT_APP_MAPBOX_TOKEN}`}
               />
 
-              <Marker interactive={false} icon={happyMapIcon} position={[-1.361876, -48.4042172]} />
+              <Marker interactive={false} icon={mapIcon} position={[-1.361876, -48.4042172]} />
             </Map>
 
             <div className="input-block">
