+++ web/src/pages/Orphanage.tsx
@@ -2,19 +2,11 @@ import React from "react";
 import { FaWhatsapp } from "react-icons/fa";
 import { FiClock, FiInfo } from "react-icons/fi";
 import { Map, Marker, TileLayer } from "react-leaflet";
-import L from 'leaflet';
 
-import mapMarkerImg from '../images/map-marker.svg';
 import '../styles/pages/orphanage.scss';
 import Sidebar from "../components/Sidebar";
+import mapIcon from "../utils/mapIcon";
 
-const happyMapIcon = L.icon({
-  iconUrl: mapMarkerImg,
-
-  iconSize: [58, 68],
-  iconAnchor: [29, 68],
-  popupAnchor: [0, -60]
-})
 
 export default function Orphanage() {
 
@@ -66,7 +58,7 @@ export default function Orphanage() {
                 <TileLayer 
                   url={`https://api.mapbox.com/styles/v1/mapbox/light-v10/tiles/256/{z}/{x}/{y}@2x?access_token=${process.env.REACT_APP_MAPBOX_TOKEN}`}
                 />
-                <Marker interactive={false} icon={happyMapIcon} position={[-1.361876, -48.4042172]} />
+                <Marker interactive={false} icon={mapIcon} position={[-1.361876, -48.4042172]} />
               </Map>
 
               <footer>
